core/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-filter==2.4.0
django-import-export==2.4.0
django-select2==7.4.2
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
Unipath==1.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
